Updated UI to pause during switch turns
Prior to this update the UI would not pause between switching players. This was fine when there was only 1 rule to adhere to, but once I added options to have multiple rules the UI needed to pause during switching players to allow the player to know why he/she lost their turn. This is currently set to happen for 1.5 seconds while disabling other buttons.