Added UI Feedback for dice rolls & logic change
Dice rolls now come with UI feedback to alert the user of whch number was rolled or if it is the next players turn. 

Also, 

Updated game logic to change the following:

When in doubles mode, rolling 2 1's resets the current score now instead of a single 1 on either die
Double 6's mode now resets the GLOBAL score when it is rolled, so back to back 6's in singles mode or double 6's in doubles mode will now reset the global score. 

After testing, I realized the doubles mode with current resets on just a single 1 was frustrating because of how often 1's were rolled when using 2 dice.